+++ packages/smarthr-ui/src/components/FormControl/index.ts
@@ -3,4 +3,6 @@ export {
   FormControl,
   labelObjectConverter,
   type ObjectLabelType,
+  SMARTHR_UI_INPUT_SELECTOR,
+  SMARTHR_UI_LABEL_TEXT_SELECTOR,
 } from './FormControl'
